modules/image_processor/requirements.txt
Pillow==9.1.1
redis==4.3.1